리팩토링
- 알람 리스트 조회 시 이미 알고 있는 유저 정보를 다시 한번 조회해서 DB에서 2번의 IO가 발생함.
  이 부분에 대한 코드를 수정하여 DB에서 IO가 1번만 발생하도록 변경함.
- delete를 할 때 JPA에서 제공하는 delete를 사용하지 않고 @Query를 통해 삭제하도록 변경함. (soft delete 사용)
- N+1 문제가 발생할 수도 있는 부분에 대해 지연로딩을 사용하고 default_batch_fetch_size를 사용해서 N+1문제를 방지함.

+++ src/main/java/com/me/sns/controller/PostController.java
@@ -56,7 +56,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(@PathVariable Integer postId) {
+    public Response<Long> likeCount(@PathVariable Integer postId) {
         return Response.success(postService.likeCount(postId));
     }
